# src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/status": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "return current status of api",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/apps": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "return list of apps",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/apps/{appName}": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "return list of tables for this app",
        "description": "",
        "parameters": [
          {
            "name": "appName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/apps/{appName}/{tableName}": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "return data content of table",
        "description": "",
        "parameters": [
          {
            "name": "appName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "tables"
        ],
        "summary": "create one item",
        "description": "",
        "parameters": [
          {
            "name": "appName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/apps/{appName}/{tableName}/{id}": {
      "get": {
        "tags": [
          "tables"
        ],
        "summary": "return one item by id",
        "description": "",
        "parameters": [
          {
            "name": "appName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "tables"
        ],
        "summary": "update one item by id",
        "description": "",
        "parameters": [
          {
            "name": "appName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "tables"
        ],
        "summary": "delete one item by id",
        "description": "",
        "parameters": [
          {
            "name": "appName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}